refactor: 💡 적절한 컴포넌트/폴더명 적용 및 불필요한 파일 삭제
✅ Closes: #5
