+++ src/components/common/Wrapper.tsx
@@ -1,13 +1,7 @@
 import { PropsWithChildren } from 'react';
-import Header from './Header';
 
 function Wrapper({ children }: PropsWithChildren) {
-  return (
-    <div className="h-full px-6">
-      <Header></Header>
-      {children}
-    </div>
-  );
+  return <div className="h-full px-6">{children}</div>;
 }
 
 export default Wrapper;
