+++ src/components/profile/FormHeader.tsx
@@ -1,7 +1,7 @@
 function AccountHeader() {
   return (
     <header>
-      <h1 className="text-title">프로필을 완성하세요</h1>
+      <h1 className="text-title-1 text-gray-50">프로필을 완성하세요</h1>
     </header>
   );
 }
