+++ src/components/profile/NoticeInfo.tsx
@@ -1,0 +1,9 @@
+function NoticeInfo() {
+  return (
+    <div>
+      <span>모든 정보는 다른 사람에게 공개돼요</span>
+    </div>
+  );
+}
+
+export default NoticeInfo;
